버튼 로딩 컴포넌트 중복 제거
